Set maxLinkRate type to uint32_t so it gets read

+++ WhateverGreen/kern_igfx.hpp
@@ -452,7 +452,7 @@ private:
 	 *
 	 *  Default value is 0x14 (5.4 Gbps, HBR2) for 4K laptop display
 	 */
-	uint8_t maxLinkRate {0x14};
+	uint32_t maxLinkRate {0x14};
 	
 	/**
 	 *  ReadAUX wrapper to modify the maximum link rate value in the DPCD buffer
